6/20 wip

+++ quests.py
@@ -1,7 +1,7 @@
 import json
 import os
 from Player import Player
-from Player import stat_checker
+from Player import stat
 
 ## Open the JSON file of inventory
 inventory = open("./inventory.json", encoding="utf8")
@@ -10,32 +10,36 @@ inventory = json.load(inventory)
 
 class NPC():
     def quest_Jack():
-        print("Quest #1 - Description: Collect 3 items, Reward: 5 XP")
+        trash = [(3), (6)]
+        print("Quest #1 - Description: Collect 3 items that are classified trash, Reward: 5 XP")
         #description
-        answer = input("Do you want to check quest status? (Y/N) ")
+        answer = input("Do you want to check quest completion status? (Y/N) ")
         if answer.upper() == "Y":
-            if int(len(inventory)) >= 3:
-                del inventory[0]
-                del inventory[0]
-                del inventory[0]
+            for i in inventory:
 
-                # Creates a new JSON file with the updated inventory
-                new_file = "updated.json"
-                with open(new_file, "w") as f:
-                    # Serialize the updated Python list to a JSON string
-                    json_string = json.dumps(inventory)
+                if int(len(inventory)) >= 3:
+                    if i['ID'] in trash:
+                        del i['Name']
+                        del inventory[0]
+                        del inventory[0]
 
-                    # Write the JSON string to the new JSON file
-                    f.write(json_string)
+                        # Creates a new JSON file with the updated inventory
+                        new_file = "updated.json"
+                        with open(new_file, "w") as f:
+                            # Serialize the updated Python list to a JSON string
+                            json_string = json.dumps(inventory)
 
-                # Overwrite the old JSON file with the new one
-                os.remove("inventory.json")
-                os.rename(new_file, "inventory.json")
+                            # Write the JSON string to the new JSON file
+                            f.write(json_string)
 
-                for i in Player:
-                    i['Total_XP'] = i['Total_XP'] + 5
-                    print("Jack: Congrats! You’ve successfully completed the quest -- Here’s your reward: *5 XP*")
-                    stat_checker()
+                        # Overwrite the old JSON file with the new one
+                        os.remove("inventory.json")
+                        os.rename(new_file, "inventory.json")
+
+                        for i in Player:
+                            i['Total_XP'] = i['Total_XP'] + 5
+                            print("Jack: Congrats! You’ve successfully completed the quest -- Here’s your reward: *5 XP*")
+                            stat.stat_checker()
             else:
                 print("Jack: Not quite. Come back when you collect more items. ")
         else:
@@ -71,7 +75,7 @@ class NPC():
                 for i in Player:
                     i['Total_XP'] = i['Total_XP'] + 10
                     print("Fiona: Nice! You finished the quest -- This is your reward: *10 XP*")
-                    stat_checker()
+                    stat.stat_checker()
             else:
                 print("Fiona: Sorry, you don't have enough items. Come back when you collect more items. ")
         else:
@@ -110,7 +114,7 @@ class NPC():
                 for i in Player:
                     i['Total_XP'] = i['Total_XP'] + 20
                     print("Winstell: Nice job! You finished the quest -- This is your reward: *20 XP*")
-                    stat_checker()
+                    stat.stat_checker()
             else:
                 print("Winstell: You haven't completed the quest yet. Come back when you collect more items. ")
         else:
